tests/requirements.txt
Django==1.10
djangorestframework==3.4.6
pytest==3.0.1
